Unify message relay between Elder and Adult
